Shadows broken on windows, GLFramebuffer Rewrite in progress

+++ app/ShatteredMoon_25/src/main.cpp
@@ -19,7 +19,7 @@ int main(int argc, char ** argv)
     glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
     glfwWindowHint(GLFW_RESIZABLE, GL_FALSE);
     float windowWidth = 1600.0f, windowHeight = 900.0f;
-    GLFWwindow * window = glfwCreateWindow(windowWidth, windowHeight, "LearnOpenGL", nullptr, nullptr);
+    GLFWwindow * window = glfwCreateWindow((int) windowWidth, (int) windowHeight, "LearnOpenGL", nullptr, nullptr);
     if (window == nullptr)
     {
         std::cout << "Failed to create GLFW window" << std::endl;
