+++ app/TestScene/src/main.cpp
@@ -18,7 +18,7 @@ int main(int argc, char ** argv)
     glfwWindowHint(GLFW_OPENGL_PROFILE, GLFW_OPENGL_CORE_PROFILE);
     glfwWindowHint(GLFW_RESIZABLE, GL_FALSE);
     float windowWidth = 1600.0f, windowHeight = 900.0f;
-    GLFWwindow * window = glfwCreateWindow(windowWidth, windowHeight, "ShatteredMoon", nullptr, nullptr);
+    GLFWwindow * window = glfwCreateWindow((int) windowWidth, (int) windowHeight, "ShatteredMoon", nullptr, nullptr);
     if (window == nullptr)
     {
         std::cout << "Failed to create GLFW window" << std::endl;
@@ -37,7 +37,7 @@ int main(int argc, char ** argv)
         std::cout << "Failed to initialize GLEW" << std::endl;
         return -1;
     }
-
+    
     Logger::SetLogLevel(WARN);
     std::shared_ptr<EngineApp> app = std::make_shared<EngineApp>(window);
     Scene * scene = new Scene();
